att basica

+++ src/models/like_model.js
@@ -1,16 +1,20 @@
-import sql from "./connect";
-import {like_count} from "./post_model";
+import sql from "./connect.js";
 
 const like_post = async (usuario_id, post_id) => {
+
+    if (!usuario_id || !post_id) {
+
+        const error = new Error('Usuário ou Post inválido');
+        error.statusCode = 400; // Bad Request
+        throw error;
+    }
+
     try {
         const query = await sql`
             INSERT INTO likes (usuario_id, postagem_id)
             VALUES (${usuario_id}, ${post_id})
             RETURNING *
         `;
-        if(query.length !== 0 || query != null) {
-            like_count(post_id, 1); //Increment like count in post_model
-        }
         return query;
     } catch (error) {
         console.error('Error liking post:', error);
@@ -19,15 +23,20 @@ const like_post = async (usuario_id, post_id) => {
 }
 
 const unlike_post = async (usuario_id, post_id) => {
+
+    if (!usuario_id || !post_id) {
+
+        const error = new Error('Usuário ou Post inválido');
+        error.statusCode = 400; // Bad Request
+        throw error;
+    }
+
     try {
         const query = await sql`
             DELETE FROM likes
             WHERE usuario_id = ${usuario_id} AND postagem_id = ${post_id}
             RETURNING *
         `;
-        if(query.length !== 0 || query != null) {
-            like_count(post_id, 0); // Decrement like count in post_model
-        }
         return query;
     } catch (error) {
         console.error('Error unliking post:', error);
@@ -36,6 +45,15 @@ const unlike_post = async (usuario_id, post_id) => {
 }
 
 const get_likes_by_post_id = async (post_id) => {
+
+    if (!post_id) {
+        const error = new Error('Usuário ou Post inválido');
+        error.statusCode = 400; // Bad Request
+        throw error;
+    }
+
+
+
     try {
         const query = await sql`
             SELECT * FROM likes
@@ -49,8 +67,34 @@ const get_likes_by_post_id = async (post_id) => {
     }
 }
 
+const get_user_liked_posts = async (user_id) => {
+
+    if (!user_id) {
+        const error = new Error('Usuário ou Post inválido');
+        error.statusCode = 400; // Bad Request
+        throw error;
+    }
+    try {
+        const query = await sql`
+            SELECT postagem_id FROM likes
+            WHERE usuario_id = ${user_id}
+            ORDER BY data DESC;
+        `;
+
+        return query;
+    } catch (error) {
+        console.error('Error fetching likes by post ID:', error);
+        throw new Error('Internal Server Error');
+    }
+
+
+}
+
+
+
 export default {
     like_post,
     unlike_post,
-    get_likes_by_post_id
+    get_likes_by_post_id,
+    get_user_liked_posts
 };
